create MySqlTwitterMapApplicaton
