+++ examples/twittermap/web/conf/routes
@@ -5,7 +5,7 @@
 # TwitterMap routes
 GET        /                                    controllers.TwitterMapApplication.index
 GET        /drugmap                             controllers.TwitterMapApplication.drugmap
-GET        /mysqlTwittermap                     controllers.TwitterMapApplication.mysqlTwittermap
+GET        /mysqlTwittermap                     controllers.MySqlTwitterMapApplication.mysqlTwittermap
 
 GET        /city/:neLat/:swLat/:neLng/:swLng    controllers.TwitterMapApplication.getCity(neLat: Double, swLat: Double, neLng: Double, swLng: Double)
